requirements.txt
fastapi
uvicorn
pytest
httpx
pydantic
pytest-cov
sqlalchemy
psycopg2-binary
alembic
python-dotenv
